+++ vm1/cpp/source/KeyboardControllerSDL.cpp
@@ -82,22 +82,22 @@ void KeyboardControllerSDL::update(SDL_Event& event)
             break;
         }
 
-        for(int i = 0; i < m_editKeys.size(); ++i) 
+        for(size_t i = 0; i < m_editKeys.size(); ++i) 
         {
             if (event.key.key == m_editKeys[i]) {
                 if(isShiftPressed){
-                    m_eventBus.publish(BankChangeEvent(i));
+                    m_eventBus.publish(BankChangeEvent(int(i)));
                 }
                 else {
-                    m_eventBus.publish(EditModeEvent(i));
+                    m_eventBus.publish(EditModeEvent(int(i)));
                 }                
                 return;
             }
         }
 
-        for(int i = 0; i < m_mediaKeys.size(); ++i) 
+        for(size_t i = 0; i < m_mediaKeys.size(); ++i) 
         {
-            int mediaSlotId = (m_registry.inputMappings().bank * MEDIA_BUTTON_COUNT) + i;
+            int mediaSlotId = (m_registry.inputMappings().bank * MEDIA_BUTTON_COUNT) + int(i);
             if (event.key.key == m_mediaKeys[i]) {
                 if(isShiftPressed){
                     m_eventBus.publish(MediaSlotEvent(mediaSlotId, false)); // do not trigger playback
